Revert "refactor: extract news uploading into VoicesUploader"
This reverts commit 5a69f0b5f9224a4584440d7ade04ecc3205c31db.

Restores uploadNews() and its createImage/queueNewsRetry helpers to
UploadManager, and puts processInBatches back as a file-private helper
in UploadManager.kt and TeamsUploader.kt.
