+++ app/src/test/java/org/ole/planet/myplanet/services/UploadManagerTest.kt
@@ -3,6 +3,7 @@ package org.ole.planet.myplanet.services
 import android.content.Context
 import android.os.SystemClock
 import android.util.Log
+import com.google.gson.Gson
 import com.google.gson.JsonObject
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -35,6 +36,8 @@ import org.ole.planet.myplanet.repository.ResourcesRepository
 import org.ole.planet.myplanet.repository.SubmissionsRepository
 import org.ole.planet.myplanet.repository.UploadRepository
 import org.ole.planet.myplanet.repository.UserRepository
+import org.ole.planet.myplanet.repository.VoicesRepository
+import org.ole.planet.myplanet.services.retry.RetryQueue
 import org.ole.planet.myplanet.services.upload.AchievementUploader
 import org.ole.planet.myplanet.services.upload.PhotoUploader
 import org.ole.planet.myplanet.services.upload.TeamsUploader
@@ -42,8 +45,8 @@ import org.ole.planet.myplanet.services.upload.UploadConfigs
 import org.ole.planet.myplanet.services.upload.UploadCoordinator
 import org.ole.planet.myplanet.services.upload.UploadError
 import org.ole.planet.myplanet.services.upload.UploadResult
-import org.ole.planet.myplanet.services.upload.VoicesUploader
 import org.ole.planet.myplanet.utils.TestDispatcherProvider
+import org.ole.planet.myplanet.utils.TestTimeProvider
 import org.ole.planet.myplanet.utils.UrlUtils
 
 @OptIn(ExperimentalCoroutinesApi::class)
@@ -57,28 +60,22 @@ class UploadManagerTest {
         coVerify(exactly = 1) { teamsUploader.uploadTeams() }
     }
 
-    @Test
-    fun `uploadNews delegates to voicesUploader`() = testScope.runTest {
-        coEvery { voicesUploader.uploadNews() } returns Unit
-        uploadManager.uploadNews()
-        advanceUntilIdle()
-        coVerify(exactly = 1) { voicesUploader.uploadNews() }
-    }
-
     private lateinit var uploadManager: UploadManager
     private val teamsUploader: TeamsUploader = mockk(relaxed = true)
     private val context: Context = mockk(relaxed = true)
     private val submissionsRepository: SubmissionsRepository = mockk(relaxed = true)
+    private val gson: Gson = mockk(relaxed = true)
     private val uploadCoordinator: UploadCoordinator = mockk(relaxed = true)
     private val uploadRepository: UploadRepository = mockk(relaxed = true)
+    private val retryQueue: RetryQueue = mockk(relaxed = true)
     private val userRepository: UserRepository = mockk(relaxed = true)
+    private val voicesRepository: VoicesRepository = mockk(relaxed = true)
     private val uploadConfigs: UploadConfigs = mockk(relaxed = true)
     private val resourcesRepository: ResourcesRepository = mockk(relaxed = true)
     private val apiInterface: ApiInterface = mockk(relaxed = true)
     private val activitiesRepository: ActivitiesRepository = mockk(relaxed = true)
     private lateinit var photoUploader: PhotoUploader
     private val achievementUploader: AchievementUploader = mockk(relaxed = true)
-    private val voicesUploader: VoicesUploader = mockk(relaxed = true)
 
     private val testDispatcher = StandardTestDispatcher()
     private val testScope = TestScope(testDispatcher)
@@ -106,10 +103,14 @@ class UploadManagerTest {
 
         uploadManager = spyk(
             UploadManager(
+                context,
                 submissionsRepository,
+                gson,
                 uploadCoordinator,
                 uploadRepository,
+                retryQueue,
                 userRepository,
+                voicesRepository,
                 uploadConfigs,
                 resourcesRepository,
                 teamsUploader,
@@ -118,7 +119,7 @@ class UploadManagerTest {
                 testScope,
                 photoUploader,
                 achievementUploader,
-                voicesUploader
+                TestTimeProvider()
             )
         )
     }
@@ -319,6 +320,61 @@ class UploadManagerTest {
         advanceUntilIdle()
 
         coVerify { listener.onSuccess("No resources to upload") }
+    }
+
+    @Test
+    fun `uploadNews derives mimeType from filename and passes to header map`() = testScope.runTest {
+        io.mockk.mockkObject(org.ole.planet.myplanet.utils.FileUtils)
+        every { org.ole.planet.myplanet.utils.FileUtils.getFileNameFromUrl("http://example.com/test_image.png") } returns "test_image.png"
+        every { org.ole.planet.myplanet.utils.FileUtils.getMimeType("test_image.png") } returns "image/png"
+
+        val imgObj = JsonObject().apply {
+            addProperty("fileName", "test_image.png")
+            addProperty("imageUrl", "http://example.com/test_image.png")
+        }
+        val imgJsonString = imgObj.toString()
+        every { gson.fromJson(imgJsonString, JsonObject::class.java) } returns imgObj
+
+        val newsJson = JsonObject().apply {
+            addProperty("message", "Hello World")
+        }
+        val newsItem = org.ole.planet.myplanet.repository.NewsUploadData(
+            id = "news1",
+            _id = "news1_id",
+            message = "Hello World",
+            imageUrls = listOf(imgJsonString),
+            newsJson = newsJson
+        )
+
+        coEvery { voicesRepository.getNewsForUpload() } returns listOf(newsItem)
+        coEvery { userRepository.getUserModel() } returns null
+
+        val imageResponseJson = JsonObject().apply {
+            addProperty("id", "res123")
+            addProperty("rev", "rev123")
+        }
+        coEvery { uploadRepository.postUpload("http://mock.url/resources", any()) } returns retrofit2.Response.success(imageResponseJson)
+
+        coEvery { uploadRepository.uploadResource(any(), any(), any()) } returns retrofit2.Response.success(JsonObject())
+
+        val bulkResponse = com.google.gson.JsonArray().apply {
+            add(JsonObject().apply {
+                addProperty("id", "news1_id")
+                addProperty("rev", "rev2")
+            })
+        }
+        coEvery { uploadRepository.postUploadArray("http://mock.url/news/_bulk_docs", any()) } returns retrofit2.Response.success(bulkResponse)
+
+        uploadManager.uploadNews()
+        advanceUntilIdle()
+
+        coVerify(exactly = 1) {
+            uploadRepository.uploadResource(
+                match { headers -> headers["Content-Type"] == "image/png" && headers["If-Match"] == "rev123" },
+                "http://mock.url/resources/res123/test_image.png",
+                any()
+            )
+        }
     }
 
     @Test
